src/azure-ai-102/language/03-genai-speech-transcribe-generate/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = ["python-dotenv", "playsound3", "azure-identity", "openai"]
